+++ src/js/casc/export-helper.js
@@ -5,6 +5,7 @@
  */
 const path = require('path');
 const util = require('util');
+const constants = require('../constants');
 const core = require('../core');
 const log = require('../log');