+++ src/js/ui/tab-models.js
@@ -275,7 +275,7 @@ const getVariantTextureIDs = (fileName) => {
 };
 
 const exportFiles = async (files, isLocal = false, exportID = -1) => {
-	const exportPaths = new FileWriter(constants.LAST_EXPORT, 'utf8');
+	const exportPaths = new FileWriter(core.view.lastExportPath, 'utf8');
 	const format = core.view.config.exportModelFormat;
 
 	const manifest = { type: 'MODELS', exportID, succeeded: [], failed: [] };